requirements.txt
discord.py==1.6.0
pony==0.7.14
requests==2.22.0